% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dahl.R
\name{dahl}
\alias{dahl}
\title{Dahl's Method for MCMC Inference}
\usage{
dahl(mcmc.object)
}
\arguments{
\item{mcmc.object}{an MCMC object

return a list:
\itemize{
\item min.sse the smallest SSE
\item cLS the index for the iteration that minimize SSE
\item cluster the final clustering outcome
}}
}
\description{
Dahl's Method for MCMC Inference
}
